requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
celery>=5.3,<6.0
redis>=5.0,<6.0
gunicorn>=22.0,<23.0
psycopg2-binary>=2.9,<3.0
whitenoise>=6.7,<7.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<0.29
